src/main/java/ni/com/prisonsolver/support/exception/RestExceptionManager.java
package ni.com.prisonsolver.support.exception;

import org.springframework.http.HttpStatus;
import org.springframework.validation.FieldError;
import org.springframework.web.bind.MethodArgumentNotValidException;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.ResponseBody;
import org.springframework.web.bind.annotation.ResponseStatus;

import java.util.List;
import java.util.Map;
import java.util.stream.Collectors;

@ControllerAdvice
public class RestExceptionManager {

  @ResponseBody
  @ExceptionHandler(NotFoundException.class)
  @ResponseStatus(HttpStatus.NOT_FOUND)
  InfoErrorMessageResponse handleNotFoundException(NotFoundException e) {
    return new InfoErrorMessageResponse(e);
  }

  @ResponseBody
  @ExceptionHandler(IllegalArgumentException.class)
  @ResponseStatus(HttpStatus.BAD_REQUEST)
  InfoErrorMessageResponse handleIllegalArgumentException(IllegalArgumentException e) {
    return new InfoErrorMessageResponse(e);
  }

  @ResponseBody
  @ExceptionHandler(MethodArgumentNotValidException.class)
  @ResponseStatus(HttpStatus.BAD_REQUEST)
  public Map<String, List<String>> handleValidationErrors(MethodArgumentNotValidException e) {
    List<String> errors = e.getBindingResult()
      .getFieldErrors()
      .stream()
      .map(FieldError::getDefaultMessage)
      .collect(Collectors.toList());
    return Map.of("errors", errors);
  }
}
